Consolidate middleware logic into route
