+++ src/lib.rs
@@ -5,7 +5,6 @@ mod cli;
 pub mod config;
 mod error;
 mod http;
-mod middleware;
 mod routes;
 mod service;
 pub mod tracing;
