+++ src/routes.rs
@@ -3,5 +3,3 @@ mod zome_call;
 
 pub use health_check::health_check;
 pub use zome_call::zome_call;
-
-pub(crate) use zome_call::PayloadQuery;
